Replace R32F texture format with R16F

+++ osu.Framework/Resources/Shaders/sh_Path.fs
@@ -24,7 +24,7 @@ layout(location = 0) out vec4 o_Colour;
 
 void main(void) 
 {
-    highp float dstFromEdge = texture(sampler2D(m_Texture, m_Sampler), v_TexCoord).r;
+    mediump float dstFromEdge = texture(sampler2D(m_Texture, m_Sampler), v_TexCoord).r;
     lowp vec4 pathCol = texture(sampler2D(m_Texture1, m_Sampler1), TexRect1.xy + vec2(dstFromEdge, 0.0) * TexRect1.zw, -0.9);
     o_Colour = getRoundedColor(vec4(pathCol.rgb, pathCol.a * float(dstFromEdge > 0.0)), v_TexCoord);
 }
